% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_csv.R
\name{build_csv}
\alias{build_csv}
\title{Exporting Timeseries CSV Files}
\usage{
build_csv(folder = "current", start.date = NULL, start.time = NULL,
  end.date = NULL, end.time = NULL, comids, regions.name = "OnionCreek")
}
\arguments{
\item{folder}{a path to a folder with netcdf files. If not yet obtained the #' Exporting Timeseries Data for NWM data for select COMIDs

This function subsets a folder of temporal .nc data to a spatial list of COMIDs and returns a .csv of time series by COMID.
If the user wants all temporal slices only the first three parameters are needed. Otherwise start.date, start.time, end.date, 
and end.time can be used to refine the data that is returned.}

\item{start.date}{in format 'YYYY-MM-DD'. If left empty the first timeseries entry will default first day in the folder.}

\item{start.time}{in format 'HH" If left empty the first time series entry will default to the first file in the folder.}

\item{end.date}{in format 'YYYY-MM-DD' If left empty the first timeseries entry will default last day in the folder.}

\item{end.time}{in format 'HH' If left empty the first time series entry will default to the last file in the folder.}

\item{regions.name}{The name of the region, this will be used to name the output files}

\item{folder}{a path to a folder with netcdf files. If not yet obtainded consider: \code{\link{get_server}}}

\item{COMID(s)}{of interest can be a single number or a list.
If not yet determined consider \code{\link{get_COMIDs}}}

\item{start.date}{in format 'YYYY-MM-DD'. If left empty the first timeseries entry will default first day in the folder.}

\item{start.time}{in format 'HH" If left empty the first time series entry will default to the first file in the folder.}

\item{end.date}{in format 'YYYY-MM-DD' If left empty the first timeseries entry will default last day in the folder.}

\item{end.time}{in format 'HH' If left empty the first time series entry will default to the last file in the folder.}

\item{COMID(s)}{of interest can be a single number or a list.
If not yet determined consider \code{\link{get_COMIDs}}}

\item{regions.name}{The name of the region, this will be used to name the output files}
}
\value{
A .csv of flow records for the given time period for each given COMID. This file will be stored in the Output subfolder and will be named according to the dates and region provided

A .csv of flow records for the given time period for each given COMID. This file will be stored in the Output subfolder and will be named according to the dates and region provided
}
\description{
This function subsets a folder of temporal .nc data to a spatial list of COMIDs and returns a .csv of time series by COMID.
If the user wants all temporal slices only the first three parameters are needed. Otherwise start.date, start.time, end.date, 
and end.time can be used to refine the data that is returned.
}
\examples{

Example 1) Return all time slices in a folder: \\cr

build_csv(folder = "F:/johnson/NWM_DATA/NetCDFs/Current",  regions.name = "Onion Creek", comids = comids)

Example 2) Return a subset of time slices in a folder \\cr 

build_csv(folder = "F:/johnson/NWM_DATA/complete2014_retro", \\cr
start.date = "2014-01-01", start.time = "00", \\cr
end.date = "2014-10-31",end.time= "23",  \\cr
regions.name = "Onion Creek", comids = comids.onion.creek)

Example 1) Return all time slices in a folder: \\cr

build_csv(folder = "F:/johnson/NWM_DATA/NetCDFs/Current",  regions.name = "Onion Creek", comids = comids)

Example 2) Return a subset of time slices in a folder \\cr 

build_csv(folder = "F:/johnson/NWM_DATA/complete2014_retro", \\cr
start.date = "2014-01-01", start.time = "00", \\cr
end.date = "2014-10-31",end.time= "23",  \\cr
regions.name = "Onion Creek", comids = comids.onion.creek)
}
\author{
Mike Johnson

Mike Johnson
}
