% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_files.R
\name{build_files}
\alias{build_files}
\title{Setting Up A File Structure}
\usage{
build_files()
}
\value{
This function will create five subfolders in your working directory set with 'setwd()'. 
These forlders are: \cr \cr
(1) Flowlines: You will need to place a shapefile of any flowlines here. \cr \cr
(2) Images:This will be where all images generated will go. \cr \cr
(3) NetCDFs: This is where all .nc files will be placed. Current forcasts obtained with the 'get_current()'
function will go in the "Current"subfolder. Output from the "get_choosen()" function will be placed in a folder created
related to its name. \cr \cr
(4) The output folder will be home to any .csv files created from the "get_CSV()" function or .gif files
generated from the "get_GIF()" function. \cr
}
\description{
This function helps define your folder strucutre to execute the NWM functions. 
Set up reuqire two things \cr \cr
(1) Define working directory using \code{\link[base]{setwd}} \cr 
(2) Execute code chunk.
}
\examples{
setwd("parent_directory")
build_files()
}
\author{
Mike Johnson and Jim Coll
}
