% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_comidList.R
\name{comidList}
\alias{comidList}
\title{Break COMIDs into sequential lists}
\usage{
comidList(AOI = NULL, nhd = NULL, comids = NULL, num)
}
\arguments{
\item{AOI}{an AOI object}

\item{num}{the number of desired lists (default 2)}
}
\value{
a list of vectors
}
\description{
Break COMIDs into sequential lists
}
\author{
Mike Johnson
}
