% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_dataDocs.R
\docType{data}
\name{comids_all}
\alias{comids_all}
\title{COMIDS}
\format{a 2729077 element \code{vector}}
\usage{
comids_all
}
\description{
Dataset containing all COMIDS in the NHDPlusV2
}
\examples{
\dontrun{
 comids_all = nwm::comids_all
}
}
\keyword{datasets}
