% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGrid_land.R
\name{define_AOI_grid}
\alias{define_AOI_grid}
\title{Get NWM grid indeces for AOI}
\usage{
define_AOI_grid(AOI)
}
\arguments{
\item{AOI}{an AOI to transform to grid}
}
\value{
a list of min/max lon/lat values and grid indeces
}
\description{
Ingests AOI object and returns grid indeces and values for the NWM native grid
}
\author{
Mike Johnson
}
