% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadNWM.R
\name{downloadNWM}
\alias{downloadNWM}
\title{Download National Water Model Data}
\usage{
downloadNWM(AOI = NULL, filelist = NULL, param = NULL)
}
\arguments{
\item{AOI}{an Area of Interest to subset to (generated with getAOI)}

\item{filelist}{a filelist generated with getFilelist}

\item{param}{a parameter to get}
}
\value{
a matrix of point values or a stack of gridded raster images
}
\description{
Download National Water Model Data
}
\author{
Mike Johnson
}
