% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_errorCheck.R
\name{error.check}
\alias{error.check}
\title{Basic Error Checking}
\usage{
error.check(error = NULL, param = NULL, config = NULL, type = NULL,
  t = NULL, f = NULL, date = NULL)
}
\arguments{
\item{error}{the type of error to check}

\item{param}{the parameter being requested}

\item{config}{the configuration of interest}

\item{type}{the output type of interest}

\item{t}{the time of forecast}

\item{f}{the time since t}

\item{date}{the date of interest}
}
\value{
a message
}
\description{
Basic Error Checking
}
