% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_usgs.R
\name{find_usgs}
\alias{find_usgs}
\title{Find USGS Gages and Respective COMIDs}
\usage{
find_usgs(flowlines_path = NULL)
}
\description{
This function can be used to indintify USGS gages within the study area as defined by the input NHD flow lines and returns a table of 
matching USGSstation, NHD comids, and their respective latitude longitude values.
}
