% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findNHD.R
\name{findNHD}
\alias{findNHD}
\title{Find NHD lines}
\usage{
findNHD(AOI, spatial = TRUE)
}
\arguments{
\item{AOI}{an Area of Interest}

\item{spatial}{spatial return object as `sp` (default is `sf`)}
}
\value{
a \code{SpatialPolylines} object
}
\description{
Find NHD lines
}
\author{
Mike Johnson
}
