% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_current.R
\name{get_current}
\alias{get_current}
\title{Getting Most Current NWM data}
\usage{
get_current()
}
\value{
This fuction saves the 18 NetCDF files related to the most recent short term forecast to the NetCDFs/Current folder.
}
\description{
This function accesses the NOAA Nomads server to download the most up to date short term forcast NetCDF files. 
These files cover the spatial domain of the continential United States. These files will be stored in the 'NetCDFs/Current' folder
created using the 'build_files()' function in the NWM package. Each time this function is run the old files will be removed.
Therefore if you see something interesting be sure to save the data elsewhere!
}
\examples{
get_current()
}
\author{
Mike Johnson
}
