% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLand_data.R
\name{get_gridded_data}
\alias{get_gridded_data}
\title{Get Land Data from a file path}
\usage{
get_gridded_data(AOI, grid, filepath = NULL, param = NULL)
}
\arguments{
\item{AOI}{an AOI to subset (generated from AOI::getAOI)}

\item{grid}{the AOI grid (generated from getGrid_land)}

\item{filepath}{filename (generated from getFiles)}

\item{param}{parameter of interest}
}
\value{
a vector of numeric values
}
\description{
Download land output from the HydroShare Thredds server from a single file.
}
\author{
Mike Johnson
}
\keyword{internal}
