% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_rating_curve.R
\name{get_rating_curves}
\alias{get_rating_curves}
\title{Get Hand Rating Curves for COMIDs of Interest}
\usage{
get_rating_curves(HUC_units, comids)
}
\description{
Get Hand Rating Curves for COMIDs of Interest
}
