% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_stage.R
\name{get_stage}
\alias{get_stage}
\title{Convert Discharge to Stage via HAND synthetic rating curves}
\usage{
get_stage(discharge_csv_path, rating_curve_path, units_discharge = "cms",
  regions.name)
}
\arguments{
\item{rating_curve_path}{The path to a rating curve .csv generated from \code{\link{get_rating_curve}}}

\item{units_discharge}{The units of the discharge values in discharge_csv_path}

\item{regions.name}{The name of the region being studied. Used for nameing the output file.}

\item{disharge_csv_path}{The path to a .csv file containing disharge data built from \code{\link{get_current}},
\code{\link{get_server}}, and \code{\link{build_csv}}.}
}
\description{
This function converts a .csv of discharge data, which can be found via \code{\link{build_csv}} to stage
using the sythetic rating curves assoicated with the HAND methodology. These rating curves can be generated via
the \code{\link{get_rating_curve}}.
}
\author{
Mike Johnson
}
