% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getChannel.R
\name{getChannel}
\alias{getChannel}
\title{Get NWM channel output}
\usage{
getChannel(AOI, filelist, param)
}
\arguments{
\item{AOI}{a AOI to subset data (generated with AOI::getAOI())}

\item{filelist}{a list of filepaths (generated with getFiles)}

\item{param}{a channel paramater}
}
\value{
a matrix of values where columns are files and rows are COMIDs
}
\description{
Download channel output from the HydroShare Thredds server for a list of input files.
Parallized version using \code{doParallel} and \code{foreach}.
}
