% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFilelist.R
\name{getFilelist}
\alias{getFilelist}
\title{Get file list}
\usage{
getFilelist(config = "medium_range", type = "channel", date = 20180615,
  t = 0, f = 3, useHTTP = FALSE)
}
\arguments{
\item{config}{model configuration}

\item{type}{output type}

\item{date}{dates of interst}

\item{t}{time of forecast}

\item{f}{hours foward from t}

\item{useHTTP}{use HTTP? default = FALSE}
}
\value{
a list of file paths
}
\description{
get a list of paths to files containing the requested data
}
\author{
Mike Johnson
}
