% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLand.R
\name{getGrid_data}
\alias{getGrid_data}
\title{Get NWM land output}
\usage{
getGrid_data(AOI, filelist, param)
}
\arguments{
\item{AOI}{a AOI to subset data (generated with AOI::getAOI())}

\item{filelist}{a list of filepaths (generated with getFiles)}

\item{param}{a channel paramater}
}
\value{
a matrix of values where columns are files and rows are COMIDs
}
\description{
Download land output from the HydroShare Thredds server for a list of input files.
Parallized version of \code{get_land_data} using \code{doParallel} and \code{foreach}.
}
