% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGrid_land.R
\name{getGrid_meta}
\alias{getGrid_meta}
\title{Get NWM grid indeces of AOI}
\usage{
getGrid_meta(AOI)
}
\arguments{
\item{AOI}{an AOI to transform to grid}
}
\value{
a list of min/max lon/lat
}
\description{
Get NWM grid indeces of AOI
}
\author{
Mike Johnson
}
