% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGridded.R
\name{getGridded}
\alias{getGridded}
\title{Get NWM gridded data output}
\usage{
getGridded(AOI, filelist, param)
}
\arguments{
\item{AOI}{a AOI to subset data (generated with AOI::getAOI())}

\item{filelist}{a list of filepaths (generated with getFiles)}

\item{param}{a channel paramater}
}
\value{
a raster stack
}
\description{
Download land output from the HydroShare Thredds server for a list of input files.
Parallized version of \code{get_land_data} using \code{doParallel} and \code{foreach}.
}
