% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRetro.R
\name{getRetro}
\alias{getRetro}
\title{Download Retrospective NWM model data}
\usage{
getRetro(date, t = 0, type = "channel", write.path = NULL)
}
\arguments{
\item{date}{a date provided as a character string in %Y%m%d}

\item{t}{the time of forecast (t)}

\item{type}{the output type}

\item{write.path}{the directory path to save to}
}
\description{
the function downloads retrospective data from the AWS server. Data is avaialable from 1993-01-01 to 2017-12-31.
All data is downlaoded to a users to disk and stored in a folder called 'retro_nwm' either in the current working directory or
in a path specified with write path.
}
\author{
Mike Johnson
}
