% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/living_flood_data.R
\name{living_flood_data}
\alias{living_flood_data}
\title{Getting Current Stage Values for Living Flood Workflow}
\usage{
living_flood_data(region.name, comid.path, rating_curve_path)
}
\arguments{
\item{region.name}{Region being modeled. Will be used to name output files}

\item{rating_curve_path}{A path to a list of rating curves produced in \code{\link{setup_livingflood()}}.  Alternatives can be generated using
\code{\link{get_rating_curves()}}}

\item{comid_path}{A path to a list of COMIDs produced in \code{\link{setup_livingflood()}}. Alternatives can be generated using
\code{\link{get_COMIDs()}}}
}
\description{
This script is part 2/2 of the living flood workflow. Once set up with \code{\link{setup_livingflood()}}
users can run this script using the outputs from \code{\link{setup_livingflood()}}. This script can be set to
run hourly constantly providing modeled stage heights and discharge values for all input COMIDs
}
\author{
Mike Johnson
}
