% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_look.R
\name{look}
\alias{look}
\title{View NWM Model Metadata}
\usage{
look(config, type = NULL)
}
\arguments{
\item{config}{\code{character} model configuration}

\item{type}{\code{character}  model type}
}
\value{
prints a description to the console
}
\description{
View NWM Model Metadata
}
\examples{
\dontrun{
look("short_range")
look("short_range", "channel")
}

}
\author{
Mike Johnson
}
