% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NWM.R
\docType{package}
\name{NWM}
\alias{NWM}
\alias{NWM-package}
\title{National Water Model Client}
\description{
\code{AOI} package
}
\details{
See the \href{https://github.com/mikejohnson51/nwm/blob/master/README.md}{README} on github
}
