% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_cidaQuery.R
\name{query_cida}
\alias{query_cida}
\title{CIDA Server Query}
\usage{
query_cida(AOI, type = "nhdflowline_network", spatial = FALSE)
}
\arguments{
\item{AOI}{and AOI object generated with the AOI package, or a bounding box}

\item{type}{the WBD or nhdplus object to return}

\item{spatial}{return object as `sp` (default is `sf`)}
}
\value{
a \code{Spatial} object
}
\description{
CIDA Server Query
}
