% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nwmMeta.R
\docType{data}
\name{reservoir}
\alias{reservoir}
\title{National Water Meta Data}
\format{An object of class \code{data.frame} with 3 rows and 3 columns.}
\usage{
reservoir
}
\description{
listed data.frame of NWM meta data
}
\author{
Mike Johnson
}
\keyword{datasets}
