% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_csv.R
\name{viz_csv}
\alias{viz_csv}
\title{Visualize Streamflow Data}
\usage{
viz_csv(type, COMIDs, csv, catchment_path, flowlines_path)
}
\arguments{
\item{type}{What type of images should be returned? Options are: \cr \cr
(1) "hydrograph" : This will return a hydrograpgh for COMID(s) across the timeseries,
 up to 25 comids can be entered at a time. A single image will be produced.

(2) "flow" : This will return a plot of the flowlines with a line width porportional
to the flow (in cfs)/10. A folder of images, one for each time step, will be produced

(3) "combo" : This will return a multiplot with both a hydrograph for a specified COMID 
and the proportional width flowlines. A folder of images, one for each time step, will be produced.}

\item{COMIDs}{requires a list of COMIDs. For type = 'hydrograph' up to 25 COMIDs can be entered. 
For type = "flow" and type = "combo" only one COMID should be entered. See \code{\link{get_COMIDS}}}

\item{csv}{path to a csv where column one is a COMIDS and column 2:X are timeseries flow values}

\item{catchment_path}{path name to NHDPlusV2 catchments shapefile (.shp)}

\item{flowlines_path}{path name to NHDPlusV2 flowlines shapefile (.shp)}
}
\value{
Images of the specified type to the "Images/" subfolder
}
\description{
This function offers three methods for visualizing streamflows from a csv \code{\link{build_csv}}. 
Options that can be returned are hydrographs, porportional symbol flowlines, and a 
combination of hydrograph and proportional flowlines. \cr \cr
}
\details{
Each time the function is executed a subfolder mirroring the csv files name will be generated in the "Images" subfolder. 
An addtion subfolder within this folder will be created named after the 'type' entered. \cr \cr
 
Each time the function is executed an additional subfolder will be created and numbered according to t
he number of times
the function has been executed for that type. \cr \cr

If all three types are executed twice the Images folder will have a subflder named after the csv file. In this folder will three subfolders: "hydrograph", "flows" and "combo".
In each of these will be two subfolders named ex: ('type'1), ('type"2).
}
\examples{

Return hydrograph

viz_csv(type = "hydrograph", COMIDs = comids[1:25], csv ="OnionCreek_20170711_20170711.csv",
catchment_path = "/Users/mikejohnson/Desktop/Tester/Flowlines/Catchments/catchments.shp",
flowlines_path = "/Users/mikejohnson/Desktop/Tester/Flowlines/Flowlines/Flowlines.shp")

Return proportionally symbolized flowlines

viz_csv(type = "flow", COMIDs = comids[56], csv ="OnionCreek_20170711_20170711.csv",
catchment_path = "/Users/mikejohnson/Desktop/Tester/Flowlines/Catchments/catchments.shp",
flowlines_path = "/Users/mikejohnson/Desktop/Tester/Flowlines/Flowlines/Flowlines.shp")

Return combontion of proportionally symbolized flowlines an hydrograph

viz_csv(type = "combo", COMIDs = comids[56], csv ="OnionCreek_20170711_20170711.csv",
catchment_path = "/Users/mikejohnson/Desktop/Tester/Flowlines/Catchments/catchments.shp",
flowlines_path = "/Users/mikejohnson/Desktop/Tester/Flowlines/Flowlines/Flowlines.shp")
}
\seealso{
\code{\link{build_csv}}

\code{\link{get_COMIDS}}
}
\author{
Mike Johnson and Jim Coll
}
