% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_current.R
\name{viz_current}
\alias{viz_current}
\title{Visualize Current Streamflow Forecasts}
\usage{
viz_current(catchment = NULL, flowlines = NULL, catchment_path = NULL,
  flowline_path = NULL, COMID, region)
}
\arguments{
\item{catchment_path}{path name to NHDPlusV2 catchments shapefile (.shp)}

\item{region}{a string of the regions name}

\item{COMIDs}{requires a single COMID. See \code{\link{get_COMIDS}}}

\item{flowlines_path}{path name to NHDPlusV2 flowlines shapefile (.shp)}
}
\value{
Images of the specified type to the "Images/" subfolder
}
\description{
Automatically generates a proportional streamflow visulization and hydrogrpah at a point.
}
\details{
Each time the function is executed the 'Current' image \cr \cr
 
Each time the function is executed an additional subfolder will be created and numbered according to t
he number of times
the function has been executed for that type. \cr \cr

If all three types are executed twice the Images folder will have a subflder named after the csv file. In this folder will three subfolders: "hydrograph", "flows" and "combo".
In each of these will be two subfolders named ex: ('type'1), ('type"2).
}
\examples{

Return hydrograph

viz_csv(type = "hydrograph", COMIDs = comids[1:25], csv ="OnionCreek_20170711_20170711.csv",
catchment_path = "/Users/mikejohnson/Desktop/Tester/Flowlines/Catchments/catchments.shp",
flowlines_path = "/Users/mikejohnson/Desktop/Tester/Flowlines/Flowlines/Flowlines.shp")

viz_current = function(catchment_path, flowlines_path, COMID, region = "Santa Barbara")
}
\seealso{
\code{\link{build_csv}}

\code{\link{get_COMIDS}}

\code{\link{get_current}}
}
\author{
Mike Johnson and Jim Coll
}
