% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vizFlows.R
\name{vizFlows}
\alias{vizFlows}
\title{Vizualize Channel Flows}
\usage{
vizFlows(AOI = NULL, nhd = NULL, data = NULL, num = 5,
  max = TRUE)
}
\arguments{
\item{AOI}{an AOI object (for background plotting) (optional if AOI is not NULL)}

\item{nhd}{an NHD set generated with getNHD (optional if AOI is not NULL)}

\item{data}{flow data data.frame}

\item{num}{number of reaches to differentiate (default = 5)}

\item{max}{if TRUE the high flows are plotted, if FALSE then low flows}
}
\description{
Vizualize Channel Flows
}
