% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_flowlines.R
\name{get_flowlines}
\alias{get_flowlines}
\title{Get NHD Flowlines}
\usage{
get_flowlines(place = NULL, area.length = 10)
}
\arguments{
\item{place}{this can be a lat,long pair, a name such as 'UCSB' or 'Austin' or a general search such as 
'the walmart near Tuscaloosa'. The API engine is driven by google via the ggmap package and thus 
will allow queries simular to Google Maps.

area.length this value defines the length of each side of the bounding box square that will be used to extract flowlines.}
}
\description{
This function takes a place name, generates a bounding box of a specified size using this 
place name as a centroid, and downloads
the contained NHDflowlines. The results are displayed in the RStudio viewer via leaflet. By hovering over each flow line
user can see the stream/river name and the respective COMID.
}
\examples{
get_flowlines("Walmart in Tuscaloosa AL")

get_flowlines("34.42 -119.69")

get_flowlines("University of Oregon")

}
\author{
Mike Johnson
}
