% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HUC_data.R
\name{get_HUC6_data}
\alias{get_HUC6_data}
\title{Get HUC6 Flowlines and Catchment Boundaries}
\usage{
get_HUC6_data(HUC6 = "010100")
}
\arguments{
\item{HUC6}{This is the six unit identifer for the HUC6 unit.}
}
\value{
This function downloads the associated shapefiles into the /Flowlines folder built using 
\code{\link{build_files}}
}
\description{
This function hits the CyberGIS FTP server to pull the flowline and catchment shapefiles for a HUC6 unit specified by 
unit code.
}
\examples{
get_HUC6_data(HUC6 = "010100")

}
\seealso{
\code{\link{build_files}}

#' @export
}
\author{
Mike Johnson and Jim Coll
}
