% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_server.R
\name{get_server}
\alias{get_server}
\title{Getting Past NWM data from HydroShare Server}
\usage{
get_server(start.date, forecast_type = "short_range", end.date = NULL,
  type = NULL, time = NULL, forecast_hour = NULL)
}
\arguments{
\item{forecast_type}{default is "short_term" but can be chaged to "medium_range"}

\item{type}{default is set to "channel' other options include 'land', 'terrain', or 'reservoir'}

\item{time}{give as "HH" this is the time the forecast is made, default set to 'ALL'}

\item{forecast_hour}{given as "HHH", this is the number of hours from the time the forecast is made (time), defualt is set to "ALL"}

\item{date}{given in "YYY-MM-DD" format}
}
\value{
This function returns all .nc files to the NetCDFs subfolder for a given day.
}
\description{
Accesses  HydroShare server to download data for a specified day. If that day is not avaiable error messages help identify days with data.
Parameters are avaliable to filter what data is downloaded. These include forecast type, model output, and time of forecast and latency. 
If left blank all files for the day will be downloaded.
}
\examples{
get_chosen(date = '2017-06-17', type = 'channel', hour = '00', forecast_hour = '000')

}
\author{
Mike Johnson and Jim Coll
}
