% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFilelist.R
\name{getFilelist}
\alias{getFilelist}
\title{Get file list}
\usage{
getFilelist(config = "short_range", type = "channel",
  startDate = NULL, endDate = NULL, t = NULL, f = NULL, n = 5,
  m = NULL)
}
\arguments{
\item{config}{model configuration}

\item{type}{output type}

\item{startDate}{dates of interst}

\item{t}{time of forecast}

\item{f}{hours foward from t}

\item{n}{if f is null n can be used to limit the number of files returned}

\item{m}{ensemble memeber (long_range only)}

\item{useHTTP}{use HTTP? default = FALSE}
}
\value{
a list of file paths
}
\description{
get a list of paths to files containing the requested data
}
\author{
Mike Johnson
}
