% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGridTime.R
\name{getGridTime}
\alias{getGridTime}
\title{Get Time for Gridded output from downloadNWM}
\usage{
getGridTime(r)
}
\arguments{
\item{r}{a raster stack}
}
\description{
Get Time for Gridded output from downloadNWM
}
