% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NWM.R, R/nwmMeta.R
\docType{package}
\name{nwm}
\alias{nwm}
\alias{nwm-package}
\alias{nwm}
\title{National Water Model Client}
\format{An object of class \code{list} of length 3.}
\usage{
nwm
}
\description{
\code{AOI} package

listed data.frame of NWM meta data
}
\details{
See the \href{https://github.com/mikejohnson51/nwm/blob/master/README.md}{README} on github
}
\author{
Mike Johnson
}
\keyword{datasets}
