% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_livingflood().R
\name{setup_livingflood}
\alias{setup_livingflood}
\title{Setting Up Data for Living Flood Workflow}
\usage{
setup_livingflood(flowlines_path, hand.raster.needed = TRUE)
}
\arguments{
\item{hand.raster.needed}{default is set to TRUE. If a user already has the necessary hand rasters then avoid long
download times by setting to FALSE.}
}
